Crate ignition_9p_wire

Core utilities for serializing and deserializing the Plan 9 Remote Resource Protocol, version 9p2000.

See the ignition-9p crate for an implementation of the 9p2000 data types and message structures built on these utilities.

Structs

BorrowedCountPrefixedList
A wrapper that borrows a slice of values and serializes them with a count prefix.

BorrowedLengthPrefixedBytes
A wrapper that borrows a slice of bytes and serializes them with a length prefix.

BorrowedSizePrefixed
A wrapper that borrows a value and serializes it with a size prefix.

LimitedWriter
Writer adaptor which limits the bytes written to an underlying writer.

OwnedCountPrefixedList
A wrapper that owns a Vec of values and (de)serializes them with a count prefix.

OwnedLengthPrefixedBytes
A wrapper that owns a Vec of bytes and (de)serializes them with a length prefix.

OwnedSizePrefixed
A wrapper that owns a value and (de)serializes it with a size prefix.

Traits

DontTouch
Provides the "don't touch" value for use in wstat requests.

EmbeddedSize
Types that embed their size as a prefix of their wire format.

ReadFrom
Types that can be deserialized from the 9p2000 wire protocol.

SerializedSize
Types that know their serialized size.

WriteTo
Types that can be serialized to the 9p2000 wire protocol.