Struct ignition_9p_wire::OwnedCountPrefixedList

pub struct OwnedCountPrefixedList<I, T> { /* fields omitted */ }

A wrapper that owns a Vec of values and (de)serializes them with a count prefix.

Efficiency note

For potentially large Vec<u8>, prefer OwnedLengthPrefixedBytes, which will make fewer calls to the underlying Read or Write.

Example

#[derive(Debug, ReadFrom, PartialEq)]
struct ExampleType(u16);

let mut data: &'static [u8] = &[2, 0x55, 0xaa, 0x34, 0x12];
assert_eq!(
    OwnedCountPrefixedList::<u8, ExampleType>::read_from(&mut data)?.into_inner().as_slice(),
    &[ExampleType(0xaa55), ExampleType(0x1234)],
);

Implementations

impl<I, T> OwnedCountPrefixedList<I, T>

pub fn new(vec: Vec<T>) -> Self

Wraps a value.

pub fn into_inner(self) -> Vec<T>

Consumes self, returning the wrapped value.