Trait ignition_9p_wire::DontTouch

pub trait DontTouch {
    fn dont_touch() -> Self;
}

Provides the "don't touch" value for use in wstat requests.

A wstat request can avoid modifying some properties of the file by providing explicit "don't touch" values in the stat data that is sent: zero-length strings for text values and the maximum unsigned value of appropriate size for integral values.

Required methods

fn dont_touch() -> Self

Returns the "don't touch" value for this type.

Loading content...