Trait ignition_9p_wire::EmbeddedSize

pub trait EmbeddedSize {
    fn embedded_size(&self) -> usize;
}

Types that embed their size as a prefix of their wire format.

The motivating example is the 9p stat struct, which embeds the length of its contents, excluding the size field itself. This trait allows stat to avoid declaring a size field and asking all users to fill it in accurately, having it automated instead.

Use this trait together with the embedded_size_prefix attribute key when deriving ReadFrom or WriteTo:

#[derive(Debug, ReadFrom, WriteTo, PartialEq)]
#[ignition_9p_wire(embedded_size_prefix = "u16")]
struct ExampleType {
    field_a: u32,
}
impl EmbeddedSize for ExampleType {
    fn embedded_size(&self) -> usize {
        4
    }
}

let mut buf: &'static [u8] = &[4, 0, 0x78, 0x56, 0x34, 0x12];
assert_eq!(
    ExampleType::read_from(&mut buf)?,
    ExampleType { field_a: 0x12345678 },
);

Required methods

fn embedded_size(&self) -> usize

Loading content...