Trait ignition_9p_wire::ReadFrom

pub trait ReadFrom: Sized {
    fn read_from<R: Read>(r: &mut R) -> Result<Self>;
}

Types that can be deserialized from the 9p2000 wire protocol.

Derivable

This trait can be used with #[derive] on structs or tuple structs if all fields are ReadFrom or are Vecs or slices of ReadFrom elements.

#[derive(ReadFrom)]
struct Example {
    field_a: u32,
    field_b: String,
}

The derived implementation reads each field in declaration order.

Attributes

A number of options are available on the #[ignition_9p_wire()] attribute to customize the behavior of a derived implementation, adding collection counts and byte sizes on the wire. All sizes are enforced during deserialization. Attempting to read past the end of a delimited region or leaving any bytes unread at the end of a delimited region raises an error.

Struct prefixes

Structs may have an embedded size prefix on the wire.

#[derive(Debug, ReadFrom, PartialEq)]
#[ignition_9p_wire(embedded_size_prefix = "u32")]
struct Example {
    field_a: u16,
    field_b: u16,
}

let mut data: &'static [u8] = &[4, 0, 0, 0, 0x55, 0xaa, 0x34, 0x12];
assert_eq!(
    Example::read_from(&mut data)?,
    Example {
        field_a: 0xaa55,
        field_b: 0x1234,
    },
);

Field prefixes

Scalar fields may be unprefixed or size-prefixed. Vec and slice fields are count-prefixed or, for u8 elements only, may be handled as length-prefixed bytes. Specifying multiple field prefix options is an error.

#[derive(Debug, ReadFrom, PartialEq)]
struct Example {
    unprefixed: u32,

    #[ignition_9p_wire(size_prefixed = "u8")]
    size_prefixed: u32,

    #[ignition_9p_wire(count_prefixed = "u16")]
    count_prefixed: Vec<u16>,

    #[ignition_9p_wire(length_prefixed_bytes = "u32")]
    length_prefixed_bytes: Vec<u8>,
}

let mut data: &'static [u8] = &[
    // unprefixed field
    0x78, 0x56, 0x34, 0x12,

    // u8-size-prefixed field
    4, 0x78, 0x56, 0x34, 0x12,

    // u16-count-prefixed field
    3, 0, 0x34, 0x12, 0x78, 0x56, 0xbc, 0x9a,

    // u32-length-prefixed bytes field
    8, 0, 0, 0, 0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
];
assert_eq!(
    Example::read_from(&mut data)?,
    Example {
        unprefixed: 0x12345678,
        size_prefixed: 0x12345678,
        count_prefixed: vec![0x1234, 0x5678, 0x9abc],
        length_prefixed_bytes: vec![0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef],
    },
);

Required methods

fn read_from<R: Read>(r: &mut R) -> Result<Self>

Loading content...