Trait ignition_9p_wire::WriteTo

pub trait WriteTo {
    fn write_to<W: Write>(&self, w: &mut W) -> Result<()>;
}

Types that can be serialized to the 9p2000 wire protocol.

Derivable

This trait can be used with #[derive] on structs or tuple structs if all fields are WriteTo or are Vecs or slices of WriteTo elements.

#[derive(WriteTo)]
struct Example {
    field_a: u32,
    field_b: String,
}

The derived implementation writes each field in declaration order.

Attributes

A number of options are available on the #[ignition_9p_wire()] attribute to customize the behavior of a derived implementation, adding collection counts and byte sizes on the wire. All sizes are enforced during serialization. Attempting to write past the end of a delimited region or leaving a delimited region less than fully written raises an error.

Struct prefixes

Structs may have an embedded size prefix on the wire. The struct must implement EmbeddedSize.

#[derive(WriteTo)]
#[ignition_9p_wire(embedded_size_prefix = "u32")]
struct Example {
    field_a: u16,
    field_b: u16,
}
impl EmbeddedSize for Example {
    fn embedded_size(&self) -> usize {
        4
    }
}

let mut data = vec![];
Example {
    field_a: 0xaa55,
    field_b: 0x1234,
}.write_to(&mut data)?;
assert_eq!(data.as_slice(), &[4, 0, 0, 0, 0x55, 0xaa, 0x34, 0x12]);

Field prefixes

Scalar fields may be unprefixed or size-prefixed. Vec and slice fields are count-prefixed or, for u8 elements only, may be handled as length-prefixed bytes. Specifying multiple field prefix options is an error.

Size-prefixed fields must implement SerializedSize.

#[derive(WriteTo)]
struct Example {
    unprefixed: u32,

    #[ignition_9p_wire(size_prefixed = "u8")]
    size_prefixed: SizedU32,

    #[ignition_9p_wire(count_prefixed = "u16")]
    count_prefixed: Vec<u16>,

    #[ignition_9p_wire(length_prefixed_bytes = "u32")]
    length_prefixed_bytes: Vec<u8>,
}
#[derive(WriteTo)]
struct SizedU32(u32);
impl SerializedSize for SizedU32 {
    fn serialized_size(&self) -> usize {
        4
    }
}

let mut data = vec![];
Example {
    unprefixed: 0x12345678,
    size_prefixed: SizedU32(0x12345678),
    count_prefixed: vec![0x1234, 0x5678, 0x9abc],
    length_prefixed_bytes: vec![0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef],
}.write_to(&mut data)?;
assert_eq!(
    data.as_slice(),
    &[
        // unprefixed field
        0x78, 0x56, 0x34, 0x12,

        // u8-size-prefixed field
        4, 0x78, 0x56, 0x34, 0x12,

        // u16-count-prefixed field
        3, 0, 0x34, 0x12, 0x78, 0x56, 0xbc, 0x9a,

        // u32-length-prefixed bytes field
        8, 0, 0, 0, 0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
    ],
);

Required methods

fn write_to<W: Write>(&self, w: &mut W) -> Result<()>

Loading content...